Try to access current tryton transaction if available
If the documentation is generated from a trytond transaction is not possible to
access proteus (and does not have any sense to do it), so we can use the current
tryton transaction to access the data.

If the transaction is not avaialbe the previous procedure using proteus is used.

As a side effect proteus is now a optional dependency.